factor type formal value out into its own library
