+++ lib/swid/ir/util/unsatisfiedTypeParameters.dart
@@ -2,6 +2,7 @@ import 'package:hydro_sdk/swid/ir/constPrimitives.dart';
 import 'package:hydro_sdk/swid/ir/swidReferenceDeclarationKind.dart';
 import 'package:hydro_sdk/swid/ir/swidType.dart';
 import 'package:hydro_sdk/swid/ir/swidTypeFormal.dart';
+import 'package:hydro_sdk/swid/ir/swidTypeFormalValue.dart';
 
 enum SwidClassTypeFormalClosureKind {
   kCloseOverTypeFormalsInStaticMembers,
@@ -19,22 +20,22 @@ List<SwidTypeFormal> unsatisfiedTypeParameters({
         fromSwidInterface: (val) => ([
           (val.referenceDeclarationKind ==
                       SwidReferenceDeclarationKind.typeParameterType &&
-                  !ancestorTypeFormals.any((x) => x.value.name == val.name)
+                  !ancestorTypeFormals.any((x) => x.value.name == val.name,)
               ? SwidTypeFormal(
-                  value: SwidTypeFormalValue.fromString(string: val.name),
+                  value: SwidTypeFormalValue.fromString(string: val.name,),
                   swidReferenceDeclarationKind:
-                      SwidReferenceDeclarationKind.typeParameterType)
+                      SwidReferenceDeclarationKind.typeParameterType,)
               : dartUnkownTypeFormal),
           ...((List<List<SwidTypeFormal>> unsatisfiedTypes) =>
               unsatisfiedTypes.isNotEmpty
                   ? unsatisfiedTypes
-                      .reduce((value, element) => [...value, ...element])
+                      .reduce((value, element) => [...value, ...element,])
                   : [])(val.typeArguments
               .map((x) => unsatisfiedTypeParameters(
                     swidType: x,
                     ancestorTypeFormals: ancestorTypeFormals,
-                  ))
-              .toList())
+                  ),)
+              .toList(),)
         ]..removeWhere((x) => x == dartUnkownTypeFormal)),
         fromSwidDefaultFormalParameter: (_) => [
           dartUnkownTypeFormal,
@@ -51,12 +52,12 @@ List<SwidTypeFormal> unsatisfiedTypeParameters({
                       ...ancestorTypeFormals,
                       ...val.typeFormals,
                     ],
-                  ))
+                  ),)
               .toList()),
           ...((List<List<SwidTypeFormal>> unsatisfiedTypes) =>
               unsatisfiedTypes.isNotEmpty
                   ? unsatisfiedTypes
-                      .reduce((value, element) => [...value, ...element])
+                      .reduce((value, element) => [...value, ...element,])
                   : [])(val.normalParameterTypes
               .map((x) => unsatisfiedTypeParameters(
                     swidType: x,
@@ -64,7 +65,7 @@ List<SwidTypeFormal> unsatisfiedTypeParameters({
                       ...ancestorTypeFormals,
                       ...val.typeFormals,
                     ],
-                  ))
+                  ),)
               .toList()),
           ...unsatisfiedTypeParameters(
             swidType: val.returnType,
@@ -78,16 +79,16 @@ List<SwidTypeFormal> unsatisfiedTypeParameters({
           ...(val.constructorType != null
               ? unsatisfiedTypeParameters(
                   swidType: SwidType.fromSwidFunctionType(
-                      swidFunctionType: val.constructorType!),
+                      swidFunctionType: val.constructorType!,),
                   ancestorTypeFormals: [
                       ...ancestorTypeFormals,
                       ...val.typeFormals,
-                    ])
+                    ],)
               : []),
           ...((List<List<SwidTypeFormal>> unsatisfiedTypes) =>
               unsatisfiedTypes.isNotEmpty
                   ? unsatisfiedTypes
-                      .reduce((value, element) => [...value, ...element])
+                      .reduce((value, element) => [...value, ...element,])
                   : [])(val.factoryConstructors
               .map((x) => unsatisfiedTypeParameters(
                       swidType: SwidType.fromSwidFunctionType(
@@ -105,7 +106,7 @@ List<SwidTypeFormal> unsatisfiedTypeParameters({
           ...((List<List<SwidTypeFormal>> unsatisfiedTypes) =>
               unsatisfiedTypes.isNotEmpty
                   ? unsatisfiedTypes
-                      .reduce((value, element) => [...value, ...element])
+                      .reduce((value, element) => [...value, ...element,])
                   : [])(val.staticMethods
               .map((x) => unsatisfiedTypeParameters(
                       swidType: SwidType.fromSwidFunctionType(
@@ -123,7 +124,7 @@ List<SwidTypeFormal> unsatisfiedTypeParameters({
           ...((List<List<SwidTypeFormal>> unsatisfiedTypes) =>
               unsatisfiedTypes.isNotEmpty
                   ? unsatisfiedTypes
-                      .reduce((value, element) => [...value, ...element])
+                      .reduce((value, element) => [...value, ...element,])
                   : [])(val.methods
               .map((x) => unsatisfiedTypeParameters(
                       swidType: SwidType.fromSwidFunctionType(
@@ -137,14 +138,14 @@ List<SwidTypeFormal> unsatisfiedTypeParameters({
           ...((List<List<SwidTypeFormal>> unsatisfiedTypes) =>
               unsatisfiedTypes.isNotEmpty
                   ? unsatisfiedTypes
-                      .reduce((value, element) => [...value, ...element])
+                      .reduce((value, element) => [...value, ...element,])
                   : [])(val.instanceFieldDeclarations.entries
               .map((x) => unsatisfiedTypeParameters(
                       swidType: x.value,
                       ancestorTypeFormals: [
                         ...ancestorTypeFormals,
                         ...val.typeFormals,
-                      ]))
+                      ],),)
               .toList()),
         ]..removeWhere((x) => x == dartUnkownTypeFormal)),
       )
