+++ test/swid/transforms/dart/rtManagedClassDeclaration/doubleLinkedQueueEntry_test.dart
@@ -1,3 +1,4 @@
+import 'package:hydro_sdk/swid/ir/swidTypeFormalValue.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:hydro_sdk/swid/backend/dart/dartRtManagedClassDeclaration.dart';
