+++ test/swid/transforms/ts/functionSignatures/positionalDelegate_test.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:hydro_sdk/swid/ir/swidTypeFormalValue.dart';
 
 import 'package:hydro_sdk/swid/backend/ts/transforms/trailingReturnTypeKind.dart';
 import 'package:hydro_sdk/swid/backend/ts/transforms/transformFunctionTypeToTs.dart';
