+++ test/swid/transforms/ts/primitiveTypeFormals_test.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:hydro_sdk/swid/ir/swidTypeFormalValue.dart';
 
 import 'package:hydro_sdk/swid/backend/ts/transforms/transformPrimitiveNamesToTs.dart';
 import 'package:hydro_sdk/swid/ir/swidClass.dart';
